Close #4 - Use Perl Artistic license
While I actually prefer the protections of the LGPL3 license, it's
really onerous and obnoxious to have to include the license info at
the head of every file, and I didn't want to work out how to get
Minilla to infer my license choice.

Bad reasons? Maybe. But LGPL continues to be a deliberate pain in the
ass, so I'm exercising my right to boycott it for shallow reasons.
And because this library isn't that important, it's not the end of
the world if I got this wrong.